+++ src/index.ts
@@ -2,8 +2,8 @@ import { login } from "@utils/loginManager";
 import { loadPlugins } from "@utils/pluginManager";
 
 async function run() {
-    await login();
-    await loadPlugins();
+  await login();
+  await loadPlugins();
 }
 
 run();