+++ src/plugin/help.ts
@@ -1,19 +1,20 @@
-import { NewMessageEvent } from "telegram/events";
 import { listCommands, getPlugin } from "@utils/pluginManager";
+import { Plugin } from "@utils/pluginBase";
+import { Api } from "telegram";
 
-const helpPlugin = {
+const helpPlugin: Plugin = {
   command: "help",
   description: "查看帮助信息",
-  commandHandler: async (event: NewMessageEvent) => {
-    const msg = event.message;
+  cmdHandler: async (msg: Api.Message) => {
     const [, ...args] = msg.message.slice(1).split(" ");
     if (args.length === 0) {
       const commands = listCommands();
       await msg.edit({
         text: `
-                可用命令列表:\n${commands
-                  .sort((a, b) => a.localeCompare(b))
-                  .join(", ")}\n使用 .help <命令> 查看具体帮助信息`,
+        可用命令列表:\n${commands
+          .sort((a, b) => a.localeCompare(b))
+          .join(", ")}\n使用 .help <命令> 查看具体帮助信息
+        `,
       });
       return;
     }