+++ src/plugin/npm.ts
@@ -1,9 +1,9 @@
-import { NewMessageEvent } from "telegram/events";
 import { Plugin } from "@utils/pluginBase";
 import { loadPlugins } from "@utils/pluginManager";
 import path from "path";
 import fs from "fs";
 import axios from "axios";
+import { Api } from "telegram";
 
 const PLUGIN_PATH = path.join(process.cwd(), "plugins");
 
@@ -12,33 +12,32 @@ async function getMediaFileName(msg: any): Promise<string> {
   return metadata.document.attributes[0].fileName;
 }
 
-async function installRemotePlugin(plugin: string, event: NewMessageEvent) {
-  await event.message.edit({ text: `正在安装插件 ${plugin}...` });
+async function installRemotePlugin(plugin: string, msg: Api.Message) {
+  await msg.edit({ text: `正在安装插件 ${plugin}...` });
   const url = `https://github.com/TeleBoxDev/TeleBox_Plugins/blob/main/plugins.json?raw=true`;
   const res = await axios.get(url);
   if (res.status === 200) {
     if (!res.data[plugin]) {
-      await event.message.edit({ text: `未找到插件 ${plugin} 的远程资源` });
+      await msg.edit({ text: `未找到插件 ${plugin} 的远程资源` });
       return;
     }
     const pluginUrl = res.data[plugin].url;
     const response = await axios.get(pluginUrl);
     if (response.status !== 200) {
-      await event.message.edit({ text: `无法下载插件 ${plugin}` });
+      await msg.edit({ text: `无法下载插件 ${plugin}` });
       return;
     }
     // 保存插件文件
     const filePath = path.join(PLUGIN_PATH, `${plugin}.ts`);
     fs.writeFileSync(filePath, response.data);
-    await event.message.edit({ text: `插件 ${plugin} 已安装并加载成功` });
+    await msg.edit({ text: `插件 ${plugin} 已安装并加载成功` });
     await loadPlugins(); // 重新加载插件
   } else {
-    await event.message.edit({ text: `无法获取远程插件库` });
+    await msg.edit({ text: `无法获取远程插件库` });
   }
 }
 
-async function installPlugin(args: string[], event: NewMessageEvent) {
-  const msg = event.message;
+async function installPlugin(args: string[], msg: Api.Message) {
   if (args.length === 1) {
     if (msg.isReply) {
       const replied = await msg.getReplyMessage();
@@ -57,27 +56,26 @@ async function installPlugin(args: string[], event: NewMessageEvent) {
     }
   } else {
     const packageName = args[1];
-    await installRemotePlugin(packageName, event);
+    await installRemotePlugin(packageName, msg);
   }
 }
 
-async function uninstallPlugin(plugin: string, event: NewMessageEvent) {
+async function uninstallPlugin(plugin: string, msg: Api.Message) {
   if (!plugin) {
-    await event.message.edit({ text: "请提供要卸载的插件名称" });
+    await msg.edit({ text: "请提供要卸载的插件名称" });
     return;
   }
   const pluginPath = path.join(PLUGIN_PATH, `${plugin}.ts`);
   if (fs.existsSync(pluginPath)) {
     fs.unlinkSync(pluginPath);
-    await event.message.edit({ text: `插件 ${plugin} 已卸载` });
+    await msg.edit({ text: `插件 ${plugin} 已卸载` });
   } else {
-    await event.message.edit({ text: `未找到插件 ${plugin}` });
+    await msg.edit({ text: `未找到插件 ${plugin}` });
   }
   await loadPlugins(); // 重新加载插件
 }
 
-async function uploadPlugin(args: string[], event: NewMessageEvent) {
-  const msg = event.message;
+async function uploadPlugin(args: string[], msg: Api.Message) {
   const pluginName = args[1];
   if (!pluginName) {
     await msg.edit({ text: "请提供插件名称" });
@@ -88,24 +86,21 @@ async function uploadPlugin(args: string[], event: NewMessageEvent) {
     await msg.edit({ text: `未找到插件 ${pluginName}` });
     return;
   }
-  await msg.client?.sendFile(
-    msg.peerId,{
+  await msg.client?.sendFile(msg.peerId, {
     file: pluginPath,
     // thumb: path.join(process.cwd(), "assets/eat/eatada.png"),
-    caption: `**TeleBox_Plugin ${pluginName} plugin.**`
-  })
+    caption: `**TeleBox_Plugin ${pluginName} plugin.**`,
+  });
 }
-
 
 const npmPlugin: Plugin = {
   command: "npm",
   description: `
-    本地资源: 对某个文件回复 npm install
-    远程资源: npm install <plugin_name> || npm i <plugin_name>
-    卸载插件: npm uninstall <plugin_name> || npm rm <plugin_name> || npm un <plugin_name> || npm remove <plugin_name>
+本地资源: 对某个文件回复 npm install
+远程资源: npm install <plugin_name> || npm i <plugin_name>
+卸载插件: npm uninstall <plugin_name> || npm rm <plugin_name> || npm un <plugin_name> || npm remove <plugin_name>
     `,
-  commandHandler: async (event: NewMessageEvent) => {
-    const msg = event.message;
+  cmdHandler: async (msg) => {
     const text = msg.message;
     const [, ...args] = text.slice(1).split(" ");
     if (args.length === 0) {
@@ -115,17 +110,16 @@ const npmPlugin: Plugin = {
 
     const cmd = args[0];
     if (cmd === "install" || cmd === "i") {
-      await installPlugin(args, event);
+      await installPlugin(args, msg);
     } else if (
       cmd === "uninstall" ||
       cmd == "un" ||
       cmd === "remove" ||
       cmd === "rm"
     ) {
-      await uninstallPlugin(args[1], event);
-    }
-    else if (cmd == "upload") {
-      await uploadPlugin(args, event);
+      await uninstallPlugin(args[1], msg);
+    } else if (cmd == "upload") {
+      await uploadPlugin(args, msg);
     }
   },
 };